requirement.txt
pikepdf>=8.0.0
Pillow>=10.0.0
ilovepdf>=1.0.0
python-dotenv>=1.0.0